refactor: avoid `tokio::spawn` for cache requests (#6229)
* refactor: avoid `tokio::spawn` for cache requests

During the happy-path, we don't need any tokio task to drive a cache
loader requests because the future issuing the request just act as a
driver. If that future is cancelled, we place the cache request in an
extra task. This avoid latencies due to task overhead and (task) context
switches for most requests. This may remove a millisecond or two from
latency but also makes the whole thing easier to analyze/profile because
we don't spawn a truckload of tasks.

This trick was borrowed from rskafka.

* refactor: split up code