feat(ingester2): persist partitions on WAL rotate
Changes the WAL rotation code to cause the current set of partitions to
be submitted for persistence.

Because rotating the WAL segment and triggering persistence is not
atomic (not under an exclusive lock preventing writes) the persisted
buffers MAY contain writes that appear in the new segment file.

In the happy/non-crash path, this will have no effect, however if the
ingester crashes and replays the WAL files, these writes will be
duplicated into object storage (where compaction will resolve the issue)
- this seems like a good trade-off, allowing us to avoid blocking write
requests for as long as it takes to rotate the WAL and mark all
partitions as persisting.

(though currently WAL files are not dropped and thus everything is
replayed all the time.)