feat(ingester2): drop WAL segments after persist
Changes the WAL rotation task to drop the WAL segment after the
partition data has completely persisted.

This logic contains a race condition outlined (at length) in the code
comments, along with a plan to resolve it once I'm back from holiday.
For now, the extremely low likelihood and minor impact of the race
occurring is likely acceptable for testing purposes.